requirements.txt
jupyter==1.0.0
numpy==1.19.5
pandas==1.3.0
scikit_learn==1.0
fastapi==0.1.0
uvicorn==0.15.0
streamlit==1.2.0
mlflow==1.22.0
psycopg2==2.9.2
python-decouple
river
